pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.77.7",
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
]
